chore: Use rimraf instead of rm -rf
Some npm scripts were using rm -rf to clean the package. Unfortunately,
Windows does not come with rm -rf built-in. rimraf provides the same
functionality as an npm package.

This change replaces rm -rf calls with rimraf and installs it as a
devDependency so that the scripts work on Windows out-of the box.