Add a LICENSE file.
Some tooling looks for a LICENSE file to be certain of the author's intent for the licensing of some code.

The package.json and README.md both state that this package is MIT licensed. I've taken the text from the link in the README (https://opensource.org/licenses/MIT) and filled in the year of the initial commit and your name.